pyproject.toml
[tool.poetry]
name = "llmgraph"
version = "0.0.4"
description = "LLM created knowledge graphs"
authors = ["Dylan Hogg <dylanhogg@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.28.1"
httpx = "^0.25.0"
networkx = "^3.1"
beautifulsoup4 = "^4.12.2"
pyvis = "^0.3.2"
joblib = "^1.3.2"
tenacity = "^8.2.3"
omegaconf = "^2.3.0"
matplotlib = "^3.8.0"
python-dotenv = "^1.0.0"
click = "^8.1.7"
typer = "^0.9.0"
rich = "^13.6.0"
tqdm = "^4.66.1"
loguru = "^0.7.2"
isort = "^5.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
black = "^23.9.1"
flake8 = "^6.1.0"

[tool.poetry.scripts]
llmgraph = 'llmgraph.console:typer_app'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
